+++ packages/kyc-controller/src/index.test.ts
@@ -12,6 +12,8 @@ describe('@metamask/kyc-controller', () => {
       alpha2ToAlpha3: expect.any(Function),
       generateKeyPair: expect.any(Function),
       decryptCredentials: expect.any(Function),
+      MoonPayFrameHandler: expect.any(Function),
+      clearMoonPaySession: expect.any(Function),
       controllerName: 'KycController',
       serviceName: 'KycService',
     });